Remove Store import in AppEffect
Removed unused import '@ngrx/store' from app.effect.ts since it was not used in the file. This change is intended to improve the readability of the codebase, streamline it, and maintain its tidiness. It’s best practice to get rid of any unused imports to enhance the overall performance.