Updated baseUrl in environment files and renamed environment.prd.ts to environment.local.ts
The baseUrl for the development environment has been changed to an empty string in environment.ts. Whereas environment.prd.ts has been renamed to environment.local.ts and its baseUrl has been set to 'https://mache-viewer-dev.4mo.co'. This is to ensure the right baseUrl is used in the right environment.