Update packages and enhance DynamoDB methods
In this commit, several changes are made to the application mainly focusing on backend and frontend package updates and enhancements to DynamoDB methods. Firstly, the frontend package '@babel/traverse' and backend packages 'lodash' and '@types/lodash' are upgraded for better performance and stability.

Secondly, improvements are made to the DynamoDB methods. In 'backend/src/infrastracture/dynamodb/dynamodb-manager.ts', 'query()', 'putAll()' and 'deleteAll()' methods are adjusted for better handling of DynamoDB operations. A new 'queryAll()' method is also added.

Lastly, updates are made to 'backend/src/infrastracture/gift.repository-impl.ts' in order to implement proper handling of gift information like checking whether the index 0 gift is valid or not, verifying if saved gift data matches with the input gift data, handling of instances where gift data is shorter than saved data etc.

The aim of these changes is to enhance the efficiency of the application by upgrading to better performing packages and refining the DynamoDB methods.